made OrbitalStrike cooler
