+++ src/main/java/apple/voltskiya/custom_mobs/heartbeat/tick/main/HighFrequencyTick.java
@@ -9,7 +9,6 @@ import java.util.Map;
 public class HighFrequencyTick implements Tickable, TickGiverable {
     private static final int TICKS_PER_TICK = 1;
     private static HighFrequencyTick instance;
-    private int currentTick = 0;
 
     private final Map<Long, Runnable> tickering = new HashMap<>();
     private long uid = 0;
@@ -20,12 +19,9 @@ public class HighFrequencyTick implements Tickable, TickGiverable {
 
     @Override
     public void tick() {
-        if (currentTick++ % TICKS_PER_TICK == 0) {
-            currentTick = 1;
-            synchronized (tickering) {
-                for (Runnable runMe : new ArrayList<>(tickering.values())) {
-                    runMe.run();
-                }
+        synchronized (tickering) {
+            for (Runnable runMe : new ArrayList<>(tickering.values())) {
+                runMe.run();
             }
         }
     }
