+++ src/main/java/apple/voltskiya/custom_mobs/heartbeat/tick/orbital_strike/large/OrbitalStrikeIndividualTicker.java
@@ -1,0 +1,105 @@
+package apple.voltskiya.custom_mobs.heartbeat.tick.orbital_strike.large;
+
+import apple.voltskiya.custom_mobs.DistanceUtils;
+import apple.voltskiya.custom_mobs.Pair;
+import apple.voltskiya.custom_mobs.heartbeat.tick.MobListSql;
+import apple.voltskiya.custom_mobs.heartbeat.tick.main.UpdatedPlayerList;
+import apple.voltskiya.custom_mobs.heartbeat.tick.orbital_strike.OrbitalStrike;
+import org.bukkit.*;
+import org.bukkit.entity.*;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.Random;
+import java.util.UUID;
+
+
+public class OrbitalStrikeIndividualTicker {
+    private final OrbitalStrikeManagerTicker.Closeness closeness;
+    private boolean isCheckStrike = false;
+    private final ArrayList<Pair<UUID, Long>> strikers = new ArrayList<>();
+    private boolean isTicking = false;
+    private long myTickerUid;
+    private final Random random = new Random();
+    private final long callerUid = UpdatedPlayerList.callerUid();
+
+
+    public OrbitalStrikeIndividualTicker(OrbitalStrikeManagerTicker.Closeness closeness) {
+        this.closeness = closeness;
+    }
+
+    public synchronized void giveStriker(Entity striker, long lastShot) {
+        this.strikers.add(new Pair<>(striker.getUniqueId(), lastShot));
+        if (!isTicking) {
+            isTicking = true;
+            this.myTickerUid = closeness.getGiver().add(this::tick);
+        }
+    }
+
+    public synchronized void tick() {
+        Iterator<Pair<UUID, Long>> strikerUidIterator = strikers.iterator();
+        boolean trim = false;
+        long now = System.currentTimeMillis();
+        while (strikerUidIterator.hasNext()) {
+            Pair<UUID, Long> strikerUid = strikerUidIterator.next();
+            if (now - strikerUid.getValue() < OrbitalStrikeManagerTicker.get().STRIKE_COOLDOWN) {
+                continue;
+            }
+            Entity striker = Bukkit.getEntity(strikerUid.getKey());
+            if (striker == null) {
+                // remove this striker D:
+                MobListSql.removeMob(strikerUid.getKey());
+                strikerUidIterator.remove();
+                trim = true;
+            } else {
+                tickStriker(striker, strikerUid);
+                if (OrbitalStrikeManagerTicker.get().amIGivingStriker(striker, closeness, strikerUid.getValue())) {
+                    MobListSql.removeMob(strikerUid.getKey());
+                    strikerUidIterator.remove();
+                    trim = true;
+                }
+            }
+        }
+        if (trim) {
+            if (isTicking && strikers.isEmpty()) {
+                closeness.getGiver().remove(myTickerUid);
+                isTicking = false;
+            }
+            strikers.trimToSize();
+        }
+    }
+
+    private synchronized void tickStriker(Entity striker, Pair<UUID, Long> strikerUid) {
+        if (isCheckStrike) {
+            if (random.nextDouble() < OrbitalStrikeManagerTicker.get().STRIKE_CHANCE * closeness.getGiver().getTickSpeed()) {
+                checkStrike(striker, strikerUid);
+            }
+        }
+    }
+
+    private synchronized void checkStrike(Entity striker, Pair<UUID, Long> strikerUid) {
+        Location strikerLocation = striker.getLocation();
+        @Nullable LivingEntity target = ((Mob) striker).getTarget();
+        if (target == null) {
+            Player closest = UpdatedPlayerList.getClosestPlayer(striker.getLocation(), callerUid);
+            if (closest != null) {
+                Location pLocation = closest.getLocation();
+                double d = DistanceUtils.distance(pLocation, strikerLocation);
+                if (d < OrbitalStrikeManagerTicker.get().STRIKE_DISTANCE && ((Mob) striker).hasLineOfSight(closest)) {
+                    target = closest;
+                }
+            }
+        }
+        if (target != null) {
+            // we have the target. time to orbital strike it
+            new OrbitalStrike(striker, target,callerUid);
+            strikerUid.setValue(System.currentTimeMillis());
+        }
+    }
+
+    public synchronized void setIsCheckStrike() {
+        this.isCheckStrike = true;
+    }
+
+}
